[Server] Extract adminOnly api route check to middleware
